backport the svc_updatefrags flood fix

+++ source/sv_send.c
@@ -649,7 +649,7 @@ SV_UpdateToReliableMessages (void)
 			host_client->sendinfo = false;
 			SV_FullClientUpdate (host_client, &sv.reliable_datagram);
 		}
-		if (host_client->old_frags != host_client->edict->v.v.frags) {
+		if (host_client->old_frags != (int) host_client->edict->v.v.frags) {
 			for (j = 0, client = svs.clients; j < MAX_CLIENTS; j++, client++) {
 				if (client->state < cs_connected)
 					continue;
